virtinst: xenbus disk index fix
Reference: bnc#872789
This is an indexing problem created by virt-manager.  It knows not
to add two IDE disks of the same name (eg, 'hda' twice) or two Xen
disks of the same name (eg, 'xvda' twice) but with the different bus
types (ide vs xen) it added xvda with hda.  These disks were then
passed to qemu where it error'ed out with the disks having the same
index (in this case both are 0).